Refactor-AdminPage :  extracted and created AdminBookings component
